chore: improve file structure
